motor driver

+++ embedded/embedded-arduino-framework/src/PID.cpp
@@ -6,20 +6,21 @@
 #define max(a, b) (((a) > (b)) ? (a) : (b))
 #define clamp(v, l, h) min(max(v, l), h)
 
-    PID::PID(const float kp,const  float ki, const float kd, const float max_output):
-            kp(kp), ki(ki), kd(kd), max_output(max_output) {}
+PID::PID(const float kp, const float ki, const float kd, const float max_output): kp(kp), ki(ki), kd(kd),
+    max_output(max_output) {
+}
 
-    float PID::get_output(const float target, const float current) {
-        const float err = target - current;
-        const auto dt = (micros() - prev_time) / 1000000.0;
+double PID::get_output(const double target, const double current) {
+    const double err = target - current;
+    const auto dt = (micros() - prev_time) / 1000000.0;
 
-        const float p = kp * err;
-        const float d = kd * (prev_err - err) / dt;
-        accum += err * dt;
-        accum = clamp(accum, -max_output, max_output);
+    const double p = kp * err;
+    const double d = kd * (prev_err - err) / dt;
+    accum += err * dt;
+    accum = clamp(accum, -max_output, max_output);
 
-        const float i = ki * accum;
+    const double i = ki * accum;
 
-        prev_err = err;
-        return p + d + i;
-    }
+    prev_err = err;
+    return p + d + i;
+}
